refactor: extract participants mapping/schema helpers into conversion_participants_mapping.py
Move 10 schema/column-mapping helper functions verbatim out of the
2,748-line conversion_participants_blueprint.py into a new sibling
module. Two of the ten (_resolve_excluded_output_columns and
_load_existing_participants_schema) are also called directly by
route-support functions that remain in the blueprint, so they're
imported back in alongside the four functions the blueprint's route
handlers use directly.
